move regional_library contents into mom_regional
